[issue-9] move rotate properties to PieceTest, remove NotrisTest
Too soon to be working on that top-level API.
